requirements.txt
ccxt==3.0.25
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.1
loguru==0.7.0
python-dotenv==1.0.0